linux-image: Remove kmod index files in prerm; always run depmod in postinst
After a crossgrade between signed and unsigned kernel packages,
purging the replaced package removes kmod index files that were
generated by the replacement package and should be retained.

The removal of kmod index files was originally implemented in
kernel-package in response to bug #66341, and has been faithfully
carried over in the maintainer scripts for src:linux.  There's no
obvious reason why they should only be removed on purge, and that bug
report suggested they could be removed earlier.  I think we can safely
remove them as soon as we know the modules are going to be removed,
i.e. at the end of the prerm script.

It is possible for an error to occur after this point, and the
postinst script will then be used to recover.  Make it run depmod in
these cases and not only if called with the argument "configure".

TODO: Check that any module packages that are being removed along
with the kernel will not run depmod again after this point.

Closes: #851695