linux-image: Only remove symlinks in postrm if the kernel image is really gone
A crossgrade between signed and unsigned kernel packages will run the
replaced package's postrm after the replacement package's postinst.
This results in removing the unversioned symlinks to the kernel and
initramfs images.

We already avoid removing the symlinks in case of an upgrade.  Change
the condition to cover crossgrades as well.