RFC: WIP: backport USB fixes (for Ampere eMAG)
John Villalovos from Ampere has backported upstream patchset to fix USB on the
Ampere eMAG devices which are part of the Debian infrastructure.

His tree can be found at: https://github.com/jvillal-amp/linux-4.19

The patches have been built and validated on top of the 4.19.132-1. It has not
been validated with 4.19.146-1, hence I would like to gather comments from the
Debian kernel team members whether these patchset can land in the Debian stable
Linux kernel. Note the patches have been submitted to the stable tree, but
those have been rejected.

Also, it is worth to comment that at this point it is probably late for the
upcoming stable release, Buster 10.6.

Signed-off-by: Héctor Orón Martínez <zumbi@debian.org>